fix(server): change code base of express server to typescript (with basic typescript config). use "MathHub\server>npm run dev" to start the server
