+++ server/README.md
@@ -1,0 +1,3 @@
+**This project use ts-node to compile typescript as javascript, please reference:**
+https://github.com/TypeStrong/ts-node
+
